Component refactor 1
