+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableCells.tsx
@@ -1,0 +1,15 @@
+import { useContext } from 'react';
+
+import { RecordTableCellsEmpty } from '@/object-record/record-table/components/RecordTableCellsEmpty';
+import { RecordTableCellsVisible } from '@/object-record/record-table/components/RecordTableCellsVisible';
+import { RecordTableRowContext } from '@/object-record/record-table/contexts/RecordTableRowContext';
+
+export const RecordTableCells = () => {
+  const { inView, isDragging } = useContext(RecordTableRowContext);
+
+  return !inView && !isDragging ? (
+    <RecordTableCellsEmpty />
+  ) : (
+    <RecordTableCellsVisible />
+  );
+};
