+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableEmpty.tsx
@@ -1,0 +1,1 @@
+export const RecordTableEmpty = () => {};
