+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableHeader.tsx
@@ -1,22 +1,20 @@
-import { useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 import { useRecoilValue } from 'recoil';
-import { IconPlus, MOBILE_VIEWPORT } from 'twenty-ui';
+import { MOBILE_VIEWPORT } from 'twenty-ui';
 
 import { RecordTableHeaderCell } from '@/object-record/record-table/components/RecordTableHeaderCell';
+import { RecordTableHeaderCheckboxColumn } from '@/object-record/record-table/components/RecordTableHeaderCheckboxColumn';
+import { RecordTableHeaderDragDropColumn } from '@/object-record/record-table/components/RecordTableHeaderDragDropColumn';
 import { useRecordTableStates } from '@/object-record/record-table/hooks/internal/useRecordTableStates';
 import { isRecordTableScrolledLeftState } from '@/object-record/record-table/states/isRecordTableScrolledLeftState';
 import { isRecordTableScrolledTopState } from '@/object-record/record-table/states/isRecordTableScrolledTopState';
-import { Dropdown } from '@/ui/layout/dropdown/components/Dropdown';
-import { useScrollWrapperScopedRef } from '@/ui/utilities/scroll/hooks/useScrollWrapperScopedRef';
-
-import { RecordTableHeaderPlusButtonContent } from './RecordTableHeaderPlusButtonContent';
-import { SelectAllCheckbox } from './SelectAllCheckbox';
 
 const StyledTableHead = styled.thead<{
   isScrolledTop?: boolean;
   isScrolledLeft?: boolean;
 }>`
+  cursor: pointer;
+
   th {
     border-block: 1px solid ${({ theme }) => theme.border.color.light};
     color: ${({ theme }) => theme.font.color.tertiary};
@@ -69,62 +67,16 @@ const StyledTableHead = styled.thead<{
     }`
         : ''}
   }
-
-  cursor: pointer;
 `;
-
-const StyledPlusIconHeaderCell = styled.th<{ isTableWiderThanScreen: boolean }>`
-  ${({ theme }) => {
-    return `
-  &:hover {
-    background: ${theme.background.transparent.light};
-  };
-  padding-left: ${theme.spacing(3)};
-  `;
-  }};
-  border-left: none !important;
-  min-width: 32px;
-  ${({ isTableWiderThanScreen, theme }) =>
-    isTableWiderThanScreen &&
-    `
-    width: 32px;
-    border-right: none !important;
-    background-color: ${theme.background.primary};
-    `};
-  z-index: 1;
-`;
-
-const StyledPlusIconContainer = styled.div`
-  align-items: center;
-  display: flex;
-  height: 32px;
-  justify-content: center;
-  width: 32px;
-`;
-
-export const HIDDEN_TABLE_COLUMN_DROPDOWN_ID =
-  'hidden-table-columns-dropdown-scope-id';
-
-const HIDDEN_TABLE_COLUMN_DROPDOWN_HOTKEY_SCOPE_ID =
-  'hidden-table-columns-dropdown-hotkey-scope-id';
 
 export const RecordTableHeader = ({
   createRecord,
 }: {
   createRecord: () => void;
 }) => {
-  const { visibleTableColumnsSelector, hiddenTableColumnsSelector } =
-    useRecordTableStates();
-
-  const scrollWrapper = useScrollWrapperScopedRef();
-  const isTableWiderThanScreen =
-    (scrollWrapper.current?.clientWidth ?? 0) <
-    (scrollWrapper.current?.scrollWidth ?? 0);
+  const { visibleTableColumnsSelector } = useRecordTableStates();
 
   const visibleTableColumns = useRecoilValue(visibleTableColumnsSelector());
-  const hiddenTableColumns = useRecoilValue(hiddenTableColumnsSelector());
-
-  const theme = useTheme();
 
   const isRecordTableScrolledTop = useRecoilValue(
     isRecordTableScrolledTopState,
@@ -141,17 +93,8 @@ export const RecordTableHeader = ({
       isScrolledLeft={isRecordTableScrolledLeft}
     >
       <tr>
-        <th></th>
-        <th
-          style={{
-            width: 30,
-            minWidth: 30,
-            maxWidth: 30,
-            borderRight: 'transparent',
-          }}
-        >
-          <SelectAllCheckbox />
-        </th>
+        <RecordTableHeaderDragDropColumn />
+        <RecordTableHeaderCheckboxColumn />
         {visibleTableColumns.map((column) => (
           <RecordTableHeaderCell
             key={column.fieldMetadataId}
@@ -159,25 +102,6 @@ export const RecordTableHeader = ({
             createRecord={createRecord}
           />
         ))}
-        <StyledPlusIconHeaderCell
-          isTableWiderThanScreen={isTableWiderThanScreen}
-        >
-          {hiddenTableColumns.length > 0 && (
-            <Dropdown
-              dropdownId={HIDDEN_TABLE_COLUMN_DROPDOWN_ID}
-              clickableComponent={
-                <StyledPlusIconContainer>
-                  <IconPlus size={theme.icon.size.md} />
-                </StyledPlusIconContainer>
-              }
-              dropdownComponents={<RecordTableHeaderPlusButtonContent />}
-              dropdownPlacement="bottom-start"
-              dropdownHotkeyScope={{
-                scope: HIDDEN_TABLE_COLUMN_DROPDOWN_HOTKEY_SCOPE_ID,
-              }}
-            />
-          )}
-        </StyledPlusIconHeaderCell>
       </tr>
     </StyledTableHead>
   );
