+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableHeaderDragDropColumn.tsx
@@ -1,0 +1,3 @@
+export const RecordTableHeaderDragDropColumn = () => {
+  return <th></th>;
+};
