+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableLastEmptyCell.tsx
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+
+import { StyledRecordTableTd } from '@/object-record/record-table/components/StyledRecordTableTd';
+import { RecordTableRowContext } from '@/object-record/record-table/contexts/RecordTableRowContext';
+
+export const RecordTableLastEmptyCell = () => {
+  const { isSelected } = useContext(RecordTableRowContext);
+
+  return <StyledRecordTableTd isSelected={isSelected} />;
+};
