+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTablePendingRow.tsx
@@ -7,7 +7,7 @@ export const RecordTablePendingRow = () => {
   const { pendingRecordIdState } = useRecordTableStates();
   const pendingRecordId = useRecoilValue(pendingRecordIdState);
 
-  if (!pendingRecordId) return;
+  if (!pendingRecordId) return <></>;
 
   return (
     <RecordTableRow
