+++ packages/twenty-front/src/modules/object-record/record-table/constants/HiddenTableColumnDropdownId.ts
@@ -1,0 +1,2 @@
+export const HIDDEN_TABLE_COLUMN_DROPDOWN_ID =
+  'hidden-table-columns-dropdown-scope-id';
