+++ packages/twenty-front/src/modules/object-record/record-table/contexts/RecordTableContext.ts
@@ -26,6 +26,8 @@ export type RecordTableContextProps = {
   onContextMenu: (event: React.MouseEvent, recordId: string) => void;
   onCellMouseEnter: (args: HandleContainerMouseEnterArgs) => void;
   visibleTableColumns: ColumnDefinition<FieldMetadata>[];
+  recordTableId: string;
+  objectNameSingular: string;
 };
 
 export const RecordTableContext = createContext<RecordTableContextProps>(
